# specs/002-canonical-source-external-id/contracts/memory-ingest.openapi.yaml
openapi: 3.1.0
info:
  title: Memory Ingest API
  version: 0.2.0
  summary: Canonical source external-id and deterministic source-id contract.
  description: >-
    Public contract for canonical/manual and direct-standard ingest after
    canonical source identity governance. Structured logs, traces, migration
    diagnostics, and bounded metrics are internal diagnostics and are not part
    of the public response schema. Public observability surface is limited to
    request-correlation headers, error-body request_id, and public provenance
    metadata under source_metadata.system.
servers:
  - url: http://127.0.0.1:3000
paths:
  /sources/register:
    post:
      summary: Register a source document and derive memory items.
      description: >-
        Canonical/manual ingest accepts only already-canonical external ids.
        Direct-standard ingest derives a canonical project-owned external id,
        preserves the original payload id as provenance metadata, and computes
        replay semantics from canonical external identity plus semantic payload
        hash.
      operationId: registerSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CanonicalRegisterRequest'
                - $ref: '#/components/schemas/OpenBadgesRegisterRequest'
                - $ref: '#/components/schemas/ClrRegisterRequest'
      responses:
        '200':
          description: Idempotent replay of an existing authoritative source.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterSourceResponse'
        '201':
          description: New source registered successfully.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterSourceResponse'
        '400':
          description: Invalid canonical identity input or unmappable supported-standard payload.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '408':
          description: Normalization timeout.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Canonical external id already exists with a semantically different payload.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload too large.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Registration write path disabled during migration execution.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sources/{source-id}:
    get:
      summary: Retrieve a source and all associated memory items.
      operationId: getSource
      parameters:
        - name: source-id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Source found.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceResponse'
        '404':
          description: Source not found.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Authoritative storage unavailable.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /memory-items/{urn}:
    get:
      summary: Retrieve a memory item by URN.
      operationId: getMemoryItem
      parameters:
        - name: urn
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Memory item found.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryItemResponse'
        '404':
          description: Memory item not found.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Authoritative storage unavailable.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /search/memory-items:
    get:
      summary: Search memory-item projections.
      description: Returns Meilisearch projection hits only. Authoritative content remains available through GET /memory-items/{urn} and GET /sources/{source-id}.
      operationId: searchMemoryItems
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: source-id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: document-type
          in: query
          required: false
          schema:
            type: string
            enum: [text, markdown, json]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Search results.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '503':
          description: Search unavailable.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /health:
    get:
      summary: Local-only liveness probe.
      operationId: getHealth
      responses:
        '200':
          description: Service is alive.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /ready:
    get:
      summary: Dependency-aware readiness probe for the write path.
      operationId: getReady
      responses:
        '200':
          description: Authoritative write path is ready.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessResponse'
        '503':
          description: Authoritative write path is unavailable.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessResponse'
components:
  headers:
    RequestIdHeader:
      description: Request correlation identifier echoed by the server.
      schema:
        type: string
  schemas:
    CanonicalRegisterRequest:
      type: object
      required:
        - title
        - external-id
        - document-type
        - content
      properties:
        title:
          type: string
          minLength: 1
        summary:
          type: string
        external-id:
          type: string
          description: Canonical project-owned source URI.
          pattern: '^https://api\.cherry-pick\.net/[a-z0-9]+/v[0-9]+p[0-9]+(?:p[0-9]+)?/[^:]+:.+$'
          examples:
            - https://api.cherry-pick.net/cc/v1p3/nebooks.co.kr:eng3-ch01
            - https://api.cherry-pick.net/qti/v3p0/kice.re.kr:20240621
        document-type:
          type: string
          enum: [text, markdown]
        content:
          type: string
        metadata:
          type: object
          additionalProperties: true
    OpenBadgesRegisterRequest:
      type: object
      required:
        - '@context'
        - type
        - id
        - name
      properties:
        '@context':
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        type:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        id:
          type: string
          description: Original standard payload identifier preserved as provenance only.
        name:
          type: string
      additionalProperties: true
    ClrRegisterRequest:
      type: object
      required:
        - '@context'
        - type
        - id
        - name
      properties:
        '@context':
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        type:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        id:
          type: string
          description: Original standard payload identifier preserved as provenance only.
        name:
          type: string
        issuer:
          type: object
          additionalProperties: true
      additionalProperties: true
    RegisterSourceResponse:
      type: object
      required:
        - source_id
        - external_id
        - document_type
        - indexing_status
        - source_metadata
        - memory_items
      properties:
        source_id:
          type: string
          format: uuid
          description: Deterministic UUID v5 internal source identifier derived from the canonical source seed.
        external_id:
          type: string
          description: Canonical project-owned source URI.
          examples:
            - https://api.cherry-pick.net/ob/v2p0/issuer.example.org:urn%3Aexample%3Abadge%3A001
        document_type:
          type: string
          enum: [text, markdown, json]
        indexing_status:
          $ref: '#/components/schemas/IndexingStatus'
        source_metadata:
          $ref: '#/components/schemas/SourceMetadataEnvelope'
        memory_items:
          type: array
          items:
            $ref: '#/components/schemas/MemoryItemSummary'
    MemoryItemSummary:
      type: object
      required:
        - urn
        - sequence
        - unit_type
      properties:
        urn:
          type: string
        sequence:
          type: integer
          minimum: 0
        unit_type:
          $ref: '#/components/schemas/MemoryUnitType'
    SourceResponse:
      type: object
      required:
        - source_id
        - external_id
        - title
        - document_type
        - source_metadata
        - memory_items
      properties:
        source_id:
          type: string
          format: uuid
          description: Deterministic UUID v5 internal source identifier derived from the canonical source seed.
        external_id:
          type: string
        title:
          type: string
        summary:
          type: string
        document_type:
          type: string
          enum: [text, markdown, json]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        indexing_status:
          $ref: '#/components/schemas/IndexingStatus'
        source_metadata:
          $ref: '#/components/schemas/SourceMetadataEnvelope'
        memory_items:
          type: array
          items:
            $ref: '#/components/schemas/MemoryItemResponse'
    MemoryItemResponse:
      type: object
      required:
        - urn
        - source_id
        - sequence
        - content
        - item_metadata
        - created_at
      properties:
        urn:
          type: string
        source_id:
          type: string
          format: uuid
        sequence:
          type: integer
          minimum: 0
        content:
          type: string
        item_metadata:
          type: object
          required:
            - unit_type
            - start_offset
            - end_offset
            - version
          properties:
            unit_type:
              $ref: '#/components/schemas/MemoryUnitType'
            start_offset:
              type: integer
              minimum: 0
            end_offset:
              type: integer
              minimum: 0
            version:
              type: string
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        source_metadata:
          type: object
          additionalProperties: true
    SearchResponse:
      type: object
      required:
        - total
        - limit
        - offset
        - items
      properties:
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchHit'
    SearchHit:
      type: object
      required:
        - urn
        - source_id
        - sequence
        - document_type
        - content_preview
      properties:
        urn:
          type: string
        source_id:
          type: string
          format: uuid
        sequence:
          type: integer
          minimum: 0
        document_type:
          type: string
          enum: [text, markdown, json]
        content_preview:
          type: string
        score:
          type: number
          format: float
    SourceMetadataEnvelope:
      type: object
      required:
        - system
      properties:
        system:
          $ref: '#/components/schemas/SourceSystemMetadata'
      additionalProperties: true
    SourceSystemMetadata:
      type: object
      description: >-
        Public provenance contract. Internal diagnostics such as raw_body_hash,
        migration_phase, legacy_resolution_path, and operator decision reasons
        are intentionally excluded from this schema.
      required:
        - canonical_id_version
        - ingest_kind
        - semantic_payload_hash
      properties:
        canonical_id_version:
          type: string
          enum: [v1]
        ingest_kind:
          type: string
          enum: [canonical, direct_standard]
        original_standard_id:
          type: string
        semantic_payload_hash:
          type: string
          description: Authoritative replay and conflict comparator.
    IndexingStatus:
      type: string
      enum: [queued, indexed, deferred]
    MemoryUnitType:
      type: string
      enum: [paragraph, section, json_document, metadata_placeholder]
    HealthResponse:
      type: object
      required:
        - status
        - components
      properties:
        status:
          type: string
          enum: [ready]
        components:
          type: object
          required:
            - service
          properties:
            service:
              $ref: '#/components/schemas/ComponentStatus'
    ReadinessResponse:
      type: object
      required:
        - status
        - components
      properties:
        status:
          type: string
          enum: [ready, degraded, down]
        components:
          type: object
          required:
            - service
            - database
            - search
          properties:
            service:
              $ref: '#/components/schemas/ComponentStatus'
            database:
              $ref: '#/components/schemas/ComponentStatus'
            search:
              $ref: '#/components/schemas/ComponentStatus'
    ComponentStatus:
      type: string
      enum: [ready, degraded, down]
    ErrorResponse:
      type: object
      required:
        - error_code
        - message
        - timestamp
        - request_id
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string